Rename package from cxcore to cxos in cx/cxos
